Cleanup external logger
